refactor(catalina/route): 요청-핸들러 매핑 관련 패키지 생성
