+++ tomcat/src/test/java/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -7,8 +7,8 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.util.Objects;
-import org.apache.catalina.DefaultDispatcher;
-import org.apache.catalina.RequestMapper;
+import org.apache.catalina.route.DefaultDispatcher;
+import org.apache.catalina.route.RequestMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
